pyproject.toml
[project]
name = "agent_demo"
version = "0.0.1"
description = "Planning, research, and report generation."
authors = [
    { name = "Xiaopeng Xu" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.1.3",
    "azure-identity>=1.21.0",
    "azure-search>=1.0.0b2",
    "azure-search-documents>=11.5.2",
    "beautifulsoup4==4.13.3",
    "duckduckgo-search>=3.0.0",
    "exa-py>=1.8.8",
    "fastapi>=0.110.0",
    "httpx>=0.24.0",
    "inquirerpy>=0.3.4",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.3",
    "json-repair>=0.7.0",
    "langchain-anthropic>=0.3.15",
    "langchain-aws>=0.2.28",
    "langchain-community>=0.3.9",
    "langchain-deepseek>=0.1.2",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-google-vertexai>=2.0.25",
    "langchain-groq>=0.2.4",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.8",
    "langchain-tavily",
    "langgraph>=0.3.5",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "langgraph-checkpoint-postgres==2.0.21",
    "langgraph-cli[inmem]>=0.3.1",
    "langsmith>=0.3.37",
    "linkup-sdk>=0.2.3",
    "litellm>=1.63.11",
    "markdownify>=0.11.6",
    "mcp>=1.9.4",
    "numpy",
    "openai>=1.99.2",
    "pandas>=2.2.3",
    "pymupdf>=1.25.3",
    "pytest",
    "python-dotenv>=1.0.1",
    "readabilipy>=0.3.0",
    "requests>=2.32.3",
    "rich>=13.0.0",
    "socksio>=1.0.0",
    "sse-starlette>=1.6.5",
    "supabase>=2.15.3",
    "tavily-python>=0.5.0",
    "uvicorn>=0.27.1",
    "wikipedia>=1.4.0",
    "xmltodict>=0.14.2",
    "yfinance>=0.2.54",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deep_research",]

[tool.setuptools.package-dir]
"deep_research" = "src/deep_research"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# for documentation
[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
extend-include = ["*.pyi"]

lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
exclude = ['^/build/']